+++ tests/test_provider_github.py
@@ -1,14 +1,15 @@
+import os
+
 import httpx
 import pytest
 
-from chatlas import ChatOpenAI
+from chatlas import ChatGithub
 
 from .conftest import (
     assert_data_extraction,
     assert_images_inline,
     assert_images_remote,
     assert_list_models,
-    assert_pdf_local,
     assert_tools_async,
     assert_tools_parallel,
     assert_tools_sequential,
@@ -18,9 +19,15 @@ from .conftest import (
     assert_turns_system,
 )
 
+api_key = os.getenv("GITHUB_TOKEN", os.getenv("GITHUB_PAT"))
+if api_key is None:
+    pytest.skip(
+        "GITHUB_TOKEN or GITHUB_PAT is not set; skipping tests", allow_module_level=True
+    )
 
-def test_openai_simple_request():
-    chat = ChatOpenAI(
+
+def test_github_simple_request():
+    chat = ChatGithub(
         system_prompt="Be as terse as possible; no punctuation",
     )
     chat.chat("What is 1 + 1?")
@@ -34,8 +41,8 @@ def test_openai_simple_request():
 
 
 @pytest.mark.asyncio
-async def test_openai_simple_streaming_request():
-    chat = ChatOpenAI(
+async def test_github_simple_streaming_request():
+    chat = ChatGithub(
         system_prompt="Be as terse as possible; no punctuation",
     )
     res = []
@@ -47,14 +54,14 @@ async def test_openai_simple_streaming_request():
     assert turn.finish_reason == "stop"
 
 
-def test_openai_respects_turns_interface():
-    chat_fun = ChatOpenAI
+def test_github_respects_turns_interface():
+    chat_fun = ChatGithub
     assert_turns_system(chat_fun)
     assert_turns_existing(chat_fun)
 
 
-def test_openai_tool_variations():
-    chat_fun = ChatOpenAI
+def test_github_tool_variations():
+    chat_fun = ChatGithub
     assert_tools_simple(chat_fun)
     assert_tools_simple_stream_content(chat_fun)
     assert_tools_parallel(chat_fun)
@@ -62,23 +69,23 @@ def test_openai_tool_variations():
 
 
 @pytest.mark.asyncio
-async def test_openai_tool_variations_async():
-    await assert_tools_async(ChatOpenAI)
+async def test_github_tool_variations_async():
+    await assert_tools_async(ChatGithub)
 
 
 def test_data_extraction():
-    assert_data_extraction(ChatOpenAI)
+    assert_data_extraction(ChatGithub)
 
 
-def test_openai_images():
-    chat_fun = ChatOpenAI
+def test_github_images():
+    chat_fun = ChatGithub
     assert_images_inline(chat_fun)
     assert_images_remote(chat_fun)
 
 
 @pytest.mark.asyncio
-async def test_openai_logprobs():
-    chat = ChatOpenAI()
+async def test_github_logprobs():
+    chat = ChatGithub()
 
     pieces = []
     async for x in await chat.stream_async("Hi", kwargs={"logprobs": True}):
@@ -93,14 +100,15 @@ async def test_openai_logprobs():
     assert len(logprobs) == len(pieces)
 
 
-def test_openai_pdf():
-    chat_fun = ChatOpenAI
-    assert_pdf_local(chat_fun)
+# Doesn't seem to be supported
+# def test_github_pdf():
+#    chat_fun = ChatGithub
+#    assert_pdf_local(chat_fun)
 
 
-def test_openai_custom_http_client():
-    ChatOpenAI(kwargs={"http_client": httpx.AsyncClient()})
+def test_github_custom_http_client():
+    ChatGithub(kwargs={"http_client": httpx.AsyncClient()})
 
 
-def test_openai_list_models():
-    assert_list_models(ChatOpenAI)
+def test_github_list_models():
+    assert_list_models(ChatGithub)